Register event listeners programmatically
This commit excludes two additional infrastructure beans that are
created by an annotation-based context: the default event listener
factory and the event listener post processor that looks up for
`@EventListener`-annotated bean methods.

The logic of creating the event listener again is abstracted with
a metadata object that is generated, similarly to the bean registration
itself.